Muutin NFmiMenuItemList luokan käyttämään NFmiPtrList luokan sijasta std::list luokkaa. Samalla laitoin dynaamiset menu-itemit std::unique_ptr luokan hallinnoimiksi. Muutin rajapintoja, jotta kaikki tarvittava voidaan toteuttaq sujuvasti std::list luokalla. Esim. Reset-Next iteraatio muutettiin begin-end tyyppiseksi, jotta saadaan käyttöön range-loopitus. Poistin NFmiMenuItemList(NFmiDrawParamList ) kontruktorin, koska sitä ei ole ikinä toteutettu ja se heitti poikkeuksen.
Nämä muutokset heijastuivat muutamiin muihin luokkiin.

+++ libs/fmi/bizcode/smetdata/include/NFmiMenuItem.h
@@ -71,7 +71,7 @@ class NFmiMenuItem
 	NFmiMenuItemList* SubMenu(void) const;
 	NFmiInfoData::Type DataType(void)const{return itsDataType;}; // 1999.09.22/Marko
 	void DataType(NFmiInfoData::Type theType){itsDataType = theType;}; // 1999.09.22/Marko
-	unsigned long NumberOfSubMenuItems(void);
+	size_t NumberOfSubMenuItems(void);
 
 	void CommandId(long theCommandId);
